AMP-5655
git-svn-id: file:///svn/trunk@8928 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/amp-test/org/dgfoundation/amp/test/orgProfile/AddNewOrgProfileWidget.java
@@ -76,8 +76,10 @@ private static Logger logger	= Logger.getLogger(AddNewOrgProfileWidget.class);
             AmpDaWidgetPlace plc=WidgetUtil.getPlace("orgprof_chart_test_place");
             if (plc != null) {
                 AmpWidgetOrgProfile wdg = (AmpWidgetOrgProfile) plc.getAssignedWidget();
-                WidgetUtil.clearPlacesForWidget(wdg.getId());
-                OrgProfileWidgetUtil.delete(wdg);
+                if (wdg != null) {
+                    WidgetUtil.clearPlacesForWidget(wdg.getId());
+                    OrgProfileWidgetUtil.delete(wdg);
+                }
                 WidgetUtil.deleteWidgetPlace(plc);
             }
         } catch (DgException ex) {
